Take care of the compiler warnings. JIRA: CXX-12888
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@99402 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/misc/xmlwrapp/errors.hpp
@@ -74,7 +74,7 @@ public:
     **/
     error_message(const std::string& message,
                   message_type msg_type,
-                  int line,
+                  long line,
                   const std::string& filename);
 
     /**
@@ -96,7 +96,7 @@ public:
      *
      * @return The line number, 0 if not available.
     **/
-    int get_line (void) const;
+    long get_line (void) const;
 
     /**
      * Get the file name.
@@ -123,7 +123,7 @@ public:
 private:
     message_type message_type_;
     std::string  message_;
-    int          line_;
+    long         line_;
     std::string  filename_;
 };
